+++ app/src/store/features/variables/slice.ts
