+++ app/src/store/features/variables/types.ts
